Migrate from unittest to pytest framework + update CICD to run locally
