+++ requirements.in
@@ -1,5 +1,7 @@
 flask-sqlalchemy
 pandas>1.4
 psycopg2-binary
+pytest
+pytest-cov
 pyyaml
 schema
