+++ requirements.txt
@@ -4,16 +4,22 @@
 #
 #    pip-compile
 #
+attrs==21.4.0
+    # via pytest
 click==8.1.3
     # via flask
 contextlib2==21.6.0
     # via schema
+coverage[toml]==6.4.2
+    # via pytest-cov
 flask==2.1.2
     # via flask-sqlalchemy
 flask-sqlalchemy==2.5.1
     # via -r requirements.in
 greenlet==1.1.2
     # via sqlalchemy
+iniconfig==1.1.1
+    # via pytest
 itsdangerous==2.1.2
     # via flask
 jinja2==3.1.2
@@ -22,9 +28,23 @@ markupsafe==2.1.1
     # via jinja2
 numpy==1.23.0
     # via pandas
+packaging==21.3
+    # via pytest
 pandas==1.4.3
     # via -r requirements.in
+pluggy==1.0.0
+    # via pytest
 psycopg2-binary==2.9.3
+    # via -r requirements.in
+py==1.11.0
+    # via pytest
+pyparsing==3.0.9
+    # via packaging
+pytest==7.1.2
+    # via
+    #   -r requirements.in
+    #   pytest-cov
+pytest-cov==3.0.0
     # via -r requirements.in
 python-dateutil==2.8.2
     # via pandas
@@ -38,5 +58,9 @@ six==1.16.0
     # via python-dateutil
 sqlalchemy==1.4.39
     # via flask-sqlalchemy
+tomli==2.0.1
+    # via
+    #   coverage
+    #   pytest
 werkzeug==2.1.2
     # via flask
