+++ tests/integration/test_stocks.py
@@ -1,34 +1,27 @@
-import unittest
-
 from schema import Schema
 
-from tests.integration.test_app import app, populate_db
 
+def test_time_series(client, populate_db):
+    """
+    GIVEN a StocksOHLCV model
+    WHEN the '/stocks/time-series/<symbol> endpoint is requested (GET)
+    THEN ensure the endpoint is up and running and the data is correctly formatted.
+    """
+    validation_schema = Schema(
+        [
+            {
+                'symbol': str,
+                'date': str,
+                'open': float,
+                'high': float,
+                'low': float,
+                'close': float,
+                'volume': int
+            }
+        ]
+    )
+    r = client.get('/stocks/time-series/AMZN')
 
-class TestStocks(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls.app = app.test_client()
-
-    def setUp(self):
-        populate_db()
-
-    def test_time_series(self):
-        """Ensure the endpoint is up and running and the data is correctly formatted."""
-        validation_schema = Schema(
-            [
-                {
-                    'symbol': str,
-                    'date': str,
-                    'open': float,
-                    'high': float,
-                    'low': float,
-                    'close': float,
-                    'volume': int
-                }
-            ]
-        )
-        r = TestStocks.app.get('/stocks/time-series/AMZN')
-
-        self.assertEqual(r.status_code, 200)
-        validation_schema.validate(r.json)
+    assert r.status_code == 200
+    assert len(r.json) == 754
+    validation_schema.validate(r.json)
